requirements.txt
tensorflow>=1.10.0
Pillow>=4.3.0
matplotlib>=2.2.3
numpy>=1.14.2
